detalied search refactored to reusable livewire component
