+++ app/Livewire/Animal.php
@@ -7,6 +7,7 @@ use App\Models\Animal as AnimalModell;
 use App\Models\Shelter;
 use App\Models\Species;
 use App\Models\Vaccine;
+use Livewire\Attributes\On;
 
 class Animal extends Component
 {
@@ -57,86 +58,27 @@ class Animal extends Component
     }
 
     /**
-     * This method is called when the user selects a species from the dropdown.
-     * It queries the database for animals of the selected species.
+     * This method redirects to the animal`s own site through the router if the user clicks on the animal card.
      */
 
-    public function updatedSelectedSpecies($name)
-    {
-        $query = AnimalModell::with('images')
-            ->whereHas('species', function ($q) use ($name) {
-                $q->where('name', $name);
-            });
-        if ($this->shelterId !== null) {
-            $query->where('shelter_id', $this->shelterId);
-        }
-        $this->animals = $query->limit($this->limit)->get();
-    }
+     #[On('animalFilterUpdated')]
+     public function filteredAnimals($animalIds)
+     {
+         $this->animals = AnimalModell::with(['images', 'shelter', 'species'])
+                                      ->whereIn('id', $animalIds)
+                                      ->get();
+     }
 
-    /**
-     * This method is responsible for the detailed filtering of animals.
+         /**
+     * This method deletes the filters and returns all the shelter`s animals
      */
 
-    public function filterAnimals(): void
-    {
-        $query = AnimalModell::with(['images', 'shelter', 'species']);
-
-        if ($this->selectedSpecies) {
-            $query->whereHas('species', function ($q) {
-                $q->where('name', $this->selectedSpecies);
-            });
-        }
-
-        if ($this->sex) {
-            $query->where('sex', $this->sex);
-        }
-
-        if ($this->age) {
-            switch ($this->age) {
-                case '<1':
-                    $query->where('age', '<', 1);
-                    break;
-                case '1-5':
-                    $query->whereBetween('age', [1, 5]);
-                    break;
-                case '>5':
-                    $query->where('age', '>', 5);
-                    break;
-            }
-        }
-
-        if (!empty($this->color)) {
-            $query->where('color', $this->color);
-        }
-
-        if (!empty($this->vaccine)) {
-            $query->whereHas('vaccines', function ($q) {
-                $q->whereIn('name', $this->vaccine);
-            });
-        }
-
-        if (!empty($this->city)) {
-            $query->whereHas('shelter', function ($q) {
-                $q->whereIn('city', $this->city);
-            });
-        }
-
-        $this->animals = $query->get();
-    }
-
-    /**
-     * This function deletes the selected filters in he detailed filters section.
-     */
-
+    #[On('filtersDeleted')]
     public function deleteFilters()
     {
-        $this->selectedSpecies = '';
-        $this->sex = null;
-        $this->age = null;
-        $this->color = [];
-        $this->vaccine = [];
-        $this->city = [];
-        $this->loadMore();
+        $this->animals = AnimalModell::with(['images', 'shelter', 'species'])
+                                     ->limit($this->limit)
+                                     ->get();
     }
 
     /**
@@ -151,7 +93,7 @@ class Animal extends Component
     /**
      * This method renders the animal cards to the dashboard site.
      */
-
+    
     public function render()
     {
         return view('livewire.animal', [
