+++ resources/views/livewire/animal.blade.php
@@ -1,6 +1,6 @@
 <x-scrollable-container>
     <h2 class="font-bold text-2xl mt-5">Animals</h2>
-    <x-animal-searchbar :species="$species" :colors="$colors" :vaccines="$vaccines" :cityes="$cityes" />
+    <livewire:AnimalFilter :species="$species" :colors="$colors" :vaccines="$vaccines" :cityes="$cityes" />
     <div x-ref="scrollContainer" x-on:scroll="scrollPos = $event.target.scrollLeft"
          class="overflow-x-auto scroll-smooth hide-scrollbar">
         <div class="grid grid-rows-2 grid-flow-col auto-cols-[250px] gap-4">
