+++ resources/views/livewire/show-shelter.blade.php
@@ -5,8 +5,8 @@
             <h3>{{$shelter->city}}</h3>
             <section class="md:flex gap-4 h-1/2">
                 <x-image-gallery :images="$shelter->images" class="w-full flex flex-col md:flex-row gap-2" />
-                    <div class="w-full md:w-1/2 h-[500px]  flex flex-col">
-                        <section class="bg-background shadow-sm rounded-xl">
+                <div class="w-full md:w-1/2 h-[500px]  flex flex-col">
+                    <section class="bg-background shadow-sm rounded-xl">
                         <h2 class="h1-style mx-2 border-b-2">{{ $shelter->name }}</h2>
                         <div class="flex items-center space-x-2">
                             <span>
@@ -45,6 +45,9 @@
                 </div>
             </section>
         </article>
+        <div class="w-4/5 mx-auto mt-8">
+            <livewire:AnimalFilter :species="$species" :colors="$colors" :vaccines="$vaccines" :cityes="$cityes" />
+        </div>
         <section class="grid grid-cols-1 md:grid-cols-4 gap-8 w-4/5 mx-auto">
             @foreach ($animals as $animal)
             <div wire:click="redirectToAnimal({{ $animal->id }})">
