+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\ProfileController;
+use App\Livewire\ListAnimals;
 use App\Livewire\ShowAnimal;
 use App\Livewire\ShowShelter;
 use App\Models\Animal;
@@ -13,6 +14,7 @@ Route::get('/', function () {
 
 Route::get('/shelters/{shelter}', ShowShelter::class)->name('shelters.show');
 Route::get('/animals/{animal}', ShowAnimal::class)->name('animal.show');
+Route::get('/animals', ListAnimals::class);
 
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
